Fdb: change type of payload sizes from int to long to prevent overflow
- change IFdbTransaction.Size, GetReadStatistics() and GetWriteStatistics() to return the size as a 64-bit integer instead of 32-bit
- prevents overflows if more than 2 GiB was processed in a single batch

+++ FoundationDB.Client/Core/IFdbTransactionHandler.cs
@@ -32,13 +32,13 @@ namespace FoundationDB.Client.Core
 	public interface IFdbTransactionHandler : IDisposable
 	{
 		/// <summary>Returns the estimated payload size of the transaction (including keys and values)</summary>
-		int Size { get; }
+		long Size { get; }
 
 		/// <summary>Returns the number of keys changed, as well as the estimated payload size</summary>
-		(int Keys, int Size) GetWriteStatistics();
+		(int Keys, long Size) GetWriteStatistics();
 
 		/// <summary>Returns the number of keys read, as well as their total size</summary>
-		(int Keys, int Size) GetReadStatistics();
+		(int Keys, long Size) GetReadStatistics();
 
 		/// <summary>Checks if this transaction handler is closed</summary>
 		bool IsClosed { get; }
